now TimeZone works as intended
